backend/requirements.txt
Django==4.2.2
djangorestframework==3.14.0
python-dotenv==1.0.0
djoser==2.2.0
Pillow==10.0.0
drf-base64==2.0
django-filter==23.2